Peg Emscripten compiler toolchain
The toolchain was accidentally pegged in CI because we cache
the target directory and we were cloning emsdk to the target
directory.

This PR busts the CircleCI cache, moves the emsdk clone to the
repo root, and introduces some more tooling for managing the
emscripten toolchain.

This PR introduces an emscriptent-toolchain file at the repo root
which acts similarly to the rust-toolchain file. The scripts added
in this PR install and activate the toolchain specified in this file.