Remove artichoke-web rubygem
`artichoke-web` exposed an API for manipulating the `window.location`
object of the webpage the playground is embedded in.

`artichoke-web` depends on `stdweb` which is unmaintained:
https://rustsec.org/advisories/RUSTSEC-2020-0056.html.

Additionally, `stdweb` does not initialize properly with the updated
emsdk toolchain this branch requires.

The `playground` depends on `stdweb` for installing an emscripten main
loop so the FFI API exposed by the `playground` stays alive. These
pieces will be replaced in subsequent commits by extern "C" bindings to
the emscripten C API.