Add an emscripten module that allows managing the main loop
Add extern "C" bindings for the emscripten C API to allow the
`playground` to install an emscripten main loop.

The playground `main` function must not terminate to keep the emscripten
runtime alive so the FFI Wasm functions for manipulating an Artichoke
interpreter continue to function.

This commit vendors an emscripten module from the MIT-licensed
https://github.com/gliheng/rust-wasm/blob/d89a71f4a68101c7b4e2944973b39a2c20b21ebe/sdl2-drag/src/emscripten.rs