Switch GitHub Pages CNAME generation from webpack to GitHub Action
Replace `cname-webpack-plugin` with the `peaceiris/actions-gh-pages`
GitHub Action used for publishing the playground in the `playground`
workflow.

This removes one npm package from the dependency tree.
`cname-webpack-plugin` does not have TypeScript declarations. Rather
than work with an "any" typed package, discontinue its use given a
suitable replacement exists.

`CNAME` is only useful for GitHub Pages deploys, so it is not necessary
to build it with webpack.