Add a default cargo-deny version to the audit CI job
The `cargo-deny` version is injected into the `audit` GitHub Actions
workflow using an organization secret. This means that PRs from GitHub
users that are not part of the @artichoke organization do not have
access to the version and the subsequent fetch of the release assets
fails.

This PR adds a fallback version for `cargo-deny` which is set if
`$version` is empty.

This change allows PRs from external contributors to build all green.