Construct GitHub Actions cache keys with hash of Cargo.lock
The existing cache key configuration was cargo culted from `async-std`
which is a library and doesn't check in its `Cargo.lock`. This mostly
worked because the playground primarily updates its deps when
`artichoke` has its git revision bumped.

`Cargo.lock` encodes the current state of dependencies, so hashing this
is sufficient to encode the current cache state.