Update rust-toolchain and emscripten-toolchain to LLVM 12
Artichoke only compiles with Rust 1.54.0. Rust 1.54.0 is based on LLVM
12 (rust-lang/rust#81451). To ensure the `emcc` compiled sources interop
with Rust code, it is important to use an `emsdk` version that is ~close
to when it was branched from LLVM 12.

According to the release notes [0], 2.0.23 - 05/26/2021 release is the
first to be based on LLVM 12.

However, Rust depends on a stubbed function in the emsdk c++ ABI that
was removed in 2.0.10. According to rust-lang/rust#85821, we might have
luck compiling with 2.0.9

[0]: https://emscripten.org/docs/introducing_emscripten/release_notes.html

emscripten-toolchain