Downgrade release profile `opt-level` from "z" to "s"
New Emscripten partially rewrites the C std lib implementation from Musl
when compiling with `-Oz` by changing the link order of some libraries.

I couldn't get this to work when `rustc` was involved, always ending up
with an unknown symbol error for `emscripten_memcpy_big` which led to
the `memcpy` builtin fataling in the resulting Wasm artifacts.

As a simple workaround, just compile with "s" (small) optimization,
which gets us most of the way there that "z" does.